Initial hand-rolled setup of Rust
A lot of things going on here:

- Use a libcore fork that disables floats. This is mainly to disable the
usage of SSE, which it seems adds a lot of boring ASM code that we want
to avoid.

- Setup a cross-compilation target for our OS, and put that information
in a way the rust compiler can understand. By some magic incantations
it's already working, but it need to be automated.

- Define language features, uses only the bare minimum from Rust proper.
This is truly low-level things.